nekoneko/agents/customer_service_agent.py
"""
宇宙旅行カスタマーサービスエージェント

このエージェントは、お客様のお問い合わせ対応を担当し、
以下の機能を提供します：

主な機能:
- 一般的な問い合わせ対応
- 旅行プランの相談
- クレーム対応
- お客様サポート
"""

from phi.agent import Agent
from phi.model.openai import OpenAIChat
from phi.tools.duckduckgo import DuckDuckGo
from pydantic import BaseModel
from typing import List, Optional
from datetime import datetime

class CustomerInquiry(BaseModel):
    inquiry_id: str
    customer_name: str
    inquiry_type: str
    content: str
    timestamp: datetime
    priority: str
    status: str

class CustomerFeedback(BaseModel):
    feedback_id: str
    booking_id: str
    rating: int
    comments: str
    timestamp: datetime
    response_required: bool

# カスタマーサービスエージェントの設定
customer_service_agent = Agent(
    name="Space Travel Customer Service Agent",
    role="Customer Service Specialist",
    model=OpenAIChat(id="gpt-4"),
    tools=[DuckDuckGo()],
    instructions=[
        "親切で丁寧な対応を心がけます",
        "お客様のご要望を正確に理解します",
        "分かりやすい説明を心がけます",
        "適切な解決策を提案します",
        "必要に応じて専門部署と連携します",
        "お客様の満足度を最優先します",
        "猫らしい温かみのある対応を心がけます"
    ],
    markdown=True
)

def handle_inquiry(inquiry: CustomerInquiry) -> str:
    """
    お客様からのお問い合わせを処理します
    
    Parameters:
        inquiry (CustomerInquiry): お問い合わせ内容
        
    Returns:
        str: 対応結果
    """
    # お問い合わせタイプに応じた処理
    response_templates = {
        "general": "にゃーご質問ありがとうございます。ご要望についてしっかりとサポートさせていただきます。",
        "booking": "にゃーご予約に関するお問い合わせありがとうございます。最適なプランをご提案させていただきます。",
        "complaint": "にゃーご不便をおかけし、大変申し訳ございません。早急に対応させていただきます。",
        "emergency": "にゃー緊急のご連絡ありがとうございます。直ちに対応チームに連絡いたします。",
        "feedback": "にゃーご意見ありがとうございます。より良いサービス提供に活かさせていただきます。"
    }
    
    base_response = response_templates.get(
        inquiry.inquiry_type,
        "にゃーお問い合わせありがとうございます。"
    )
    
    # 優先度に応じた対応時間の設定
    sla_hours = {
        "high": 1,
        "medium": 4,
        "low": 24
    }
    
    response_time = sla_hours.get(inquiry.priority, 24)
    
    return f"{base_response}\n\n{response_time}時間以内に詳細な回答を差し上げます。"

def process_feedback(feedback: CustomerFeedback) -> None:
    """
    お客様からのフィードバックを処理します
    
    Parameters:
        feedback (CustomerFeedback): フィードバック内容
    """
    # 評価に応じたアクション
    if feedback.rating <= 3:
        # 低評価の場合は即時対応
        if feedback.response_required:
            # フォローアップチームに通知
            send_followup_notification(feedback)
    
    # フィードバックの分析と保存
    analyze_and_store_feedback(feedback)

def send_followup_notification(feedback: CustomerFeedback) -> None:
    """
    フォローアップチームに通知を送信します
    """
    # 通知処理の実装
    pass

def analyze_and_store_feedback(feedback: CustomerFeedback) -> None:
    """
    フィードバックの分析と保存を行います
    """
    # 分析と保存処理の実装
    pass

if __name__ == "__main__":
    # エージェントのテスト実行
    customer_service_agent.print_response(
        """
        初めて宇宙旅行を検討しています。
        どのような準備が必要か、また費用の目安を
        教えていただけますでしょうか？
        """,
        stream=True
    )
